Extract EmbeddedCommonModule.
